fix(windows): Finalize adaptive log trust boundaries
- order adaptive apply events solely by the globally increasing encoder epoch while retaining the session epoch only for correlation and failing closed on allocation overflow

- commit the coalescing slot replacement and monotonic watermark under one nonblocking ownership boundary so stale, contended, and disconnected drops cannot advance admission state

- create current and rotated service log lineage with an explicit protected LocalSystem-owned descriptor and semantically validate or safely recreate existing paths

- remove the unsupported Windows NSIS and CPack service installation path so the protected MSI directory contract is the only supported installer behavior

- report stalled service event workers to the Windows Event Log before deadline-based detachment while preserving bounded shutdown

- cover smaller replacement session epochs, lock contention, disconnected publication, secure file creation, worker abandonment, and MSI-only packaging contracts

+++ engine/lumen-host/src/platform/windows/native_video.rs
@@ -2,7 +2,7 @@ use std::ffi::c_void;
 use std::mem::ManuallyDrop;
 use std::ptr;
 use std::slice;
-use std::sync::atomic::{AtomicU32, Ordering};
+use std::sync::atomic::{AtomicU64, Ordering};
 use std::sync::{mpsc, Arc, Mutex};
 use std::thread;
 use std::time::{Duration, Instant};
@@ -56,7 +56,7 @@ pub(super) struct NativeMediaFoundation {
     adaptive_event_publisher: Arc<dyn WindowsServiceEventPublisher>,
     sessions: AbandonableSessionSlot<NativeMediaFoundationSession>,
     retired_workers: RetiredWorkerRegistry,
-    next_encoder_epoch: AtomicU32,
+    next_encoder_epoch: AtomicU64,
 }
 
 struct NativeMediaFoundationSession {
@@ -143,7 +143,7 @@ enum NativeMediaFoundationCommand {
 #[derive(Clone, Copy, Debug, Eq, PartialEq)]
 struct NativeVideoEncoderPlan {
     session_epoch: u32,
-    encoder_epoch: u32,
+    encoder_epoch: u64,
     codec: PlatformVideoCodec,
     width: u32,
     height: u32,
@@ -163,7 +163,7 @@ impl NativeMediaFoundation {
             adaptive_event_publisher,
             sessions: AbandonableSessionSlot::default(),
             retired_workers: RetiredWorkerRegistry::new(MAXIMUM_RETIRED_MEDIA_FOUNDATION_WORKERS),
-            next_encoder_epoch: AtomicU32::new(1),
+            next_encoder_epoch: AtomicU64::new(1),
         }
     }
